Optimise line height computation (#449)
This optimises line height computation by pre-computing it for each
"run" in "push_run" rather than computing for each glyph after line
breaking. The performance gains are modest (but non-zero) compared to
main.

I would like someone else to confirm that this optimisation is valid,
but it looks to me like we're already relying on the invariant of "one
set of styles per run" in this part of the code for other purposes.

---------

Signed-off-by: Nico Burns <nico@nicoburns.com>