fontique: Make the yeslogic-fontconfig-sys/dlopen feature optional
The problem is that yeslogic-fontconfig-sys is actually not following
the cargo best practice to have the feature as additive, and the dlopen
feature changes the API of the crate.
Thus, if you have in your dependency tree crates that uses yeslogic-fontconfig-sys
with the dlopen feature and some that do not, it breaks the crates
that depends on it without enabling the feature.

This makes the dlopen an optional feature of fontique.
It is disabled by default because the feature set of fontique without
fontconfig doesn't make much sense.

Linking directly against the fontcinfig library makes the dependency
explicit and allow tool to figure out extra dependency easily.

This should therefore also solve issue #464