Fix crash for empty layout (#493)
With latest Parley `main`, Parley is crashing in the empty layout case
(`text = ""`). There is code to add a dummy space (not ideal seeing as
this actually gets added to text inputs - but that was a pre-existing
issue), but Parley now expects there to be at least one style.

Without this fix, Parley crashes on the `panic!("No style info");`
below.

An alternative fix (that I think would also work) would be to ensure
that the `TreeBuilder` always pushes at least one style span (even in
the empty case). Let me know if you prefer that and want me to try it.

---------

Signed-off-by: Nico Burns <nico@nicoburns.com>